Introduce `postgresql.conf` processing
Previously, script expected a bunch of pre-existing locations for
different parts of the system. This made it troublesome to use script
in the pre-existing environment (a bunch of symlinks was necessary).

Introduce one more column in the `pgtab` -- PG_CONF, that points
to the `postgresql.conf` to be used. All other locations will be
obtained from the configuration.

Improve error reporting and a couple of default values on the way.